Refactor: Separate function to module. #69
